Split integration tests into more files
