+++ test/integration/escape-all/bash.test.js
@@ -1,5 +1,5 @@
 /**
- * @overview Contains integration tests for `shescape.quote` for the
+ * @overview Contains integration tests for `shescape.escapeAll` for the
  * Bourne-again shell (Bash).
  * @license MIT
  */
@@ -8,14 +8,14 @@ import test from "ava";
 
 import { constants, generate } from "../_.js";
 
-import { quote } from "shescape";
+import { escapeAll } from "shescape";
 
 const runTest = constants.isWindows ? test.skip : test;
 
 runTest(`input is escaped for ${constants.binBash}`, (t) => {
-  for (const scenario of generate.quoteExamples(constants.binBash)) {
+  for (const scenario of generate.escapeExamples(constants.binBash)) {
     const { expected, input, options } = scenario;
-    const result = quote(input, options);
-    t.is(result, expected);
+    const result = escapeAll([input], options);
+    t.deepEqual(result, [expected]);
   }
 });
