+++ test/integration/escape-all/zsh.test.js
@@ -1,5 +1,5 @@
 /**
- * @overview Contains integration tests for `shescape.escape` for the Z shell
+ * @overview Contains integration tests for `shescape.escapeAll` for the Z shell
  * (Zsh).
  * @license MIT
  */
@@ -8,14 +8,14 @@ import test from "ava";
 
 import { constants, generate } from "../_.js";
 
-import { escape } from "shescape";
+import { escapeAll } from "shescape";
 
 const runTest = constants.isWindows ? test.skip : test;
 
 runTest(`input is escaped for ${constants.binZsh}`, (t) => {
   for (const scenario of generate.escapeExamples(constants.binZsh)) {
     const { expected, input, options } = scenario;
-    const result = escape(input, options);
-    t.is(result, expected);
+    const result = escapeAll([input], options);
+    t.deepEqual(result, [expected]);
   }
 });
