+++ test/integration/quote-all/invalid.test.js
@@ -1,5 +1,5 @@
 /**
- * @overview Contains integration tests for invalid use of `shescape.escape`.
+ * @overview Contains integration tests for invalid use of `shescape.quote`.
  * @license MIT
  */
 
@@ -8,14 +8,15 @@ import test from "ava";
 
 import { arbitrary, constants, macros } from "../_.js";
 
-import { escape } from "shescape";
+import { quoteAll } from "shescape";
 
 testProp("invalid arguments", [arbitrary.shescapeOptions()], (t, options) => {
   for (const { value } of constants.illegalArguments) {
-    t.throws(() => escape(value, options), { instanceOf: TypeError });
+    t.throws(() => quoteAll([value], options), { instanceOf: TypeError });
+    t.throws(() => quoteAll(value, options), { instanceOf: TypeError });
   }
 });
 
 test(macros.prototypePollution, (_, payload) => {
-  escape("a", payload);
+  quoteAll(["a"], payload);
 });
