+++ test/integration/quote-all/powershell-exe.test.js
@@ -1,5 +1,5 @@
 /**
- * @overview Contains integration tests for `shescape.escape` for Windows
+ * @overview Contains integration tests for `shescape.quoteAll` for Windows
  * PowerShell.
  * @license MIT
  */
@@ -8,14 +8,14 @@ import test from "ava";
 
 import { constants, generate } from "../_.js";
 
-import { escape } from "shescape";
+import { quoteAll } from "shescape";
 
 const runTest = constants.isWindows ? test : test.skip;
 
 runTest(`input is escaped for ${constants.binPowerShell}`, (t) => {
-  for (const scenario of generate.escapeExamples(constants.binPowerShell)) {
+  for (const scenario of generate.quoteExamples(constants.binPowerShell)) {
     const { expected, input, options } = scenario;
-    const result = escape(input, options);
-    t.is(result, expected);
+    const result = quoteAll([input], options);
+    t.deepEqual(result, [expected]);
   }
 });
