+++ test/integration/quote/commonjs.test.js
@@ -1,6 +1,6 @@
 /**
  * @overview Contains integration tests for the CommonJS version of
- * `shescape.escape`.
+ * `shescape.quote`.
  * @license MIT
  */
 
@@ -8,15 +8,15 @@ import { testProp } from "@fast-check/ava";
 
 import { arbitrary } from "../_.js";
 
-import { escape } from "shescape";
-import { escape as escapeCjs } from "../../../index.cjs";
+import { quote } from "shescape";
+import { quote as quoteCjs } from "../../../index.cjs";
 
 testProp(
   "esm === cjs",
   [arbitrary.shescapeArg(), arbitrary.shescapeOptions()],
   (t, arg, options) => {
-    const resultEsm = escape(arg, options);
-    const resultCjs = escapeCjs(arg, options);
+    const resultEsm = quote(arg, options);
+    const resultCjs = quoteCjs(arg, options);
     t.is(resultEsm, resultCjs);
   },
 );
