+++ test/integration/quote/valid.test.js
@@ -1,5 +1,5 @@
 /**
- * @overview Contains integration tests for valid use of `shescape.escape`.
+ * @overview Contains integration tests for valid use of `shescape.quote`.
  * @license MIT
  */
 
@@ -7,13 +7,13 @@ import { testProp } from "@fast-check/ava";
 
 import { arbitrary } from "../_.js";
 
-import { escape } from "shescape";
+import { quote } from "shescape";
 
 testProp(
-  "return values",
+  "return value",
   [arbitrary.shescapeArg(), arbitrary.shescapeOptions()],
   (t, arg, options) => {
-    const result = escape(arg, options);
+    const result = quote(arg, options);
     t.is(typeof result, "string");
   },
 );
